Added rectangular tube.

+++ scripts/tests.py
@@ -62,15 +62,21 @@ def compare_images(a, b, c):
 
 def update_image(tmp_name, png_name):
     """Update an image only if different, otherwise just change the mod time"""
+    # print("in update_image\n")
+    # print("tmp name =", tmp_name, "\n")
+    # print("png name =", png_name, "\n")
     diff_name = png_name.replace('.png', '_diff.png')
     pixels = compare_images(png_name, tmp_name, diff_name)
     if pixels < 0 or pixels > threshold:
-        shutil.copyfile(tmp_name, png_name)
         print(Fore.YELLOW + png_name + " updated" + Fore.WHITE, pixels if pixels > 0 else '')
+        if os.path.isfile(tmp_name):
+            shutil.copyfile(tmp_name, png_name)
     else:
         os.utime(png_name, None)
         os.remove(diff_name)
-    os.remove(tmp_name)
+    if os.path.isfile(tmp_name):
+        os.remove(tmp_name)
+    print("\n")
 
 
 def depluralise(name):
